requirements-dev.txt
sphinx
tox
pytest-cov
flake8
flake8-import-order
mypy